+++ R/as_index.R
@@ -73,20 +73,19 @@ as_index.default <- function(x, ...) {
 #' @export
 as_index.matrix <- function(x, ..., chainable = TRUE, contrib = FALSE) {
   chkDots(...)
-  levels <- rownames(x) %||% seq_len(nrow(x))
-  periods <- colnames(x) %||% seq_len(ncol(x))
+  levels <- as.character(rownames(x) %||% seq_len(nrow(x)))
+  periods <- as.character(colnames(x) %||% seq_len(ncol(x)))
 
-  index <- index_skeleton(levels, periods)
-  for (t in seq_along(periods)) {
-    index[[t]] <- as.numeric(x[, t])
-  }
+  index <- as.numeric(x)
+  dim(index) <- c(length(levels), length(periods))
 
-  if (contrib) {
-    contributions <- index2contrib(index, levels, periods)
-  } else {
-    contributions <- contrib_skeleton(levels, periods)
-  }
-  piar_index(index, contributions, levels, periods, chainable)
+  piar_index(
+    index,
+    if (contrib) index2contrib(index, levels, periods),
+    levels,
+    periods,
+    chainable = chainable
+  )
 }
 
 #' @rdname as_index
@@ -100,28 +99,19 @@ as_index.data.frame <- function(x, ..., contrib = FALSE) {
   x[1:2] <- lapply(x[1:2], as.factor)
   time <- levels(x[[1L]])
   levels <- levels(x[[2L]])
-  # elementary_index() usually gives NaN for missing cells.
-  index <- matrix(
-    NA_real_,
-    nrow = length(levels),
-    ncol = length(time),
-    dimnames = list(levels, time)
-  )
+  index <- index_skeleton(levels, time)
+  dimnames(index) <- list(levels, time)
   index[as.matrix(x[2:1])] <- as.numeric(x[[3L]])
 
   if (contrib && length(x) > 3L) {
-    contributions <- matrix(
-      list(numeric(0L)),
-      nrow = length(levels),
-      ncol = length(time),
-      dimnames = list(levels, time)
-    )
+    contributions <- contrib_skeleton(levels, time)
+    dimnames(contributions) <- list(levels, time)
     contributions[as.matrix(x[2:1])] <- x[[4L]]
 
     index <- as_index(index, ...)
-    for (t in seq_along(time)) {
-      index$contrib[[t]][] <- lapply(contributions[, t], valid_contrib)
-    }
+    # FIXME: I don't like adding the contributions after making the index.
+    index$contrib <- lapply(contributions, valid_contrib)
+    dim(index$contrib) <- dim(contributions)
   } else {
     index <- as_index(index, contrib = contrib, ...)
   }