+++ R/update.aggregation_structure.R
@@ -52,18 +52,22 @@
 update.piar_aggregation_structure <- function(
   object,
   index,
-  period = end(index),
   ...,
+  period = NULL,
   r = 1
 ) {
   chkDots(...)
   price_update <- gpindex::factor_weights(r)
-  index <- chain(as_index(index))
-  period <- match_time(as.character(period), index)
+  index <- as_index(index, chainable = FALSE)
+  period <- if (!is.null(period)) {
+    match_time(as.character(period), index)
+  } else {
+    ntime(index)
+  }
   eas <- match_eas(object, index)
   if (anyNA(eas)) {
     warning("not all weights in 'object' have a corresponding index value")
   }
-  weights(object) <- price_update(index$index[[period]][eas], object$weights)
+  weights(object) <- price_update(index$index[, period][eas], object$weights)
   object
 }