+++ NEWS.md
@@ -1,3 +1,57 @@
+# piar (development version)
+
+## Significant changes
+
+- The internal representation of index objects has changed so that both
+the `index` and `contrib` components of an index are now matrices instead of
+lists. This simplifies the code base and in some cases improves performance, but
+will break any code that directly uses the internal components of an index.
+
+- A new internal optimization means that index objects without contributions now
+consume about half as much memory.
+
+- `contrib()` and `contrib2DF()` are no longer generic functions. (They were
+originally generic to allow for the possibility of an index subclass without
+contributions, but that is no longer required.)
+
+- In most cases methods have optional arguments after the `...`. This means that
+in some cases arguments need to be named that previously could be passes by position.
+
+   - `link` and `base` in `chain()`, `unchain()`, and `rebase()`.
+   - `period` in `update()`.
+   
+- The default behavior for aggregating duplicate contributions is to sum them.
+This means that it is now possible to have duplicate products across levels
+in `elementary_index()`.
+
+- The `margin` argument in `split(index)` has been renamed to `along`.
+
+## Improvements
+
+- Added a new function `combine_classifications()` to more easily build an
+aggregation structure that spans multiple hierarchical classifications.
+
+- Subscripting an index object with a matrix is now more flexible and can
+return/replace a list of index objects.
+
+- Added a new function `impute_prices()` to consolidate the different
+price-imputation functions.
+
+## Bug fixes
+
+- Setting `contrib = FALSE` in `aggregate(index)` always returns an index with
+no contributions.
+
+- `set_weights()` no longer returns its result invisibly.
+
+- `update(pias, index)` now correctly coerces `index` to a fixed-base index when
+it is not an index object, as documented.
+
+## Deprecations
+
+- `carry_forward()`, `carry_backward()`, and `shadow_price()` are deprecated.
+Use `impute_prices()` instead.
+
 # piar 0.8.3
 
 ## Improvements